re-structure namespace condense read fields and add enum value check and exception
